+++ tests/test_events/handler/handleCreate.test.js
@@ -4,9 +4,9 @@ import vscode from "vscode";
 import untildify from "untildify";
 import getBranchName from "current-git-branch";
 
-import {readYML} from "../../../src/utils/common";
-import {DEFAULT_BRANCH, DIFF_SOURCE} from "../../../src/constants";
+import {DEFAULT_BRANCH} from "../../../src/constants";
 import {
+    assertNewFileEvent,
     Config,
     getConfigFilePath,
     getSyncIgnoreFilePath,
@@ -15,6 +15,7 @@ import {
 } from "../../helpers/helpers";
 import {pathUtils} from "../../../src/utils/path_utils";
 import {eventHandler} from "../../../src/events/event_handler";
+import {populateBuffer} from "../../../src/codesyncd/populate_buffer";
 
 
 describe("handleNewFile",  () => {
@@ -40,11 +41,12 @@ describe("handleNewFile",  () => {
     const shadowRepoBranchPath = pathUtilsObj.getShadowRepoBranchPath();
     const originalsRepoBranchPath = pathUtilsObj.getOriginalsRepoBranchPath();
 
-    const newFilePath = path.join(repoPath, "new.js");
+    const newRelPath = "new.js";
+    const newFilePath = path.join(repoPath, newRelPath);
     const newDirectoryPath = path.join(repoPath, "new");
     const syncIgnorePath = getSyncIgnoreFilePath(repoPath);
-    const shadowFilePath = path.join(shadowRepoBranchPath, "new.js");
-    const originalsFilePath = path.join(originalsRepoBranchPath, "new.js");
+    const shadowFilePath = path.join(shadowRepoBranchPath, newRelPath);
+    const originalsFilePath = path.join(originalsRepoBranchPath, newRelPath);
     const syncIgnoreData = ".git\n\n\n.skip_repo_1\nignore.js";
 
     const ignorableFilePath = path.join(repoPath, "ignore.js");
@@ -118,21 +120,7 @@ describe("handleNewFile",  () => {
         };
         handler.handleCreateEvent(event);
         // Verify file has been created in the .shadow repo and .originals repos
-        expect(fs.existsSync(shadowFilePath)).toBe(true);
-        expect(fs.existsSync(originalsFilePath)).toBe(true);
-        // Verify correct diff file has been generated
-        let diffFiles = fs.readdirSync(diffsRepo);
-        expect(diffFiles).toHaveLength(1);
-        const diffFilePath = path.join(diffsRepo, diffFiles[0]);
-        const diffData = readYML(diffFilePath);
-        expect(diffData.source).toEqual(DIFF_SOURCE);
-        expect(diffData.is_new_file).toBe(true);
-        expect(diffData.is_rename).toBeFalsy();
-        expect(diffData.is_deleted).toBeFalsy();
-        expect(diffData.repo_path).toEqual(repoPath);
-        expect(diffData.branch).toEqual(DEFAULT_BRANCH);
-        expect(diffData.file_relative_path).toEqual("new.js");
-        expect(diffData.diff).toEqual("");
+        expect(assertNewFileEvent(repoPath, newRelPath)).toBe(true);
     });
 
     test("Event: handlePastedFile, Repo not synced", () => {
@@ -151,43 +139,20 @@ describe("handleNewFile",  () => {
     test("Event: handlePastedFile, Synced Repo", () => {
         const handler = new eventHandler();
         handler.handlePastedFile(newFilePath);
-        // Verify file has been created in the .shadow repo and .originals repos
-        expect(fs.existsSync(shadowFilePath)).toBe(true);
-        expect(fs.existsSync(originalsFilePath)).toBe(true);
-        // Verify correct diff file has been generated
-        let diffFiles = fs.readdirSync(diffsRepo);
-        expect(diffFiles).toHaveLength(1);
-        const diffFilePath = path.join(diffsRepo, diffFiles[0]);
-        const diffData = readYML(diffFilePath);
-        expect(diffData.source).toEqual(DIFF_SOURCE);
-        expect(diffData.is_new_file).toBe(true);
-        expect(diffData.is_rename).toBeFalsy();
-        expect(diffData.is_deleted).toBeFalsy();
-        expect(diffData.repo_path).toEqual(repoPath);
-        expect(diffData.branch).toEqual(DEFAULT_BRANCH);
-        expect(diffData.file_relative_path).toEqual("new.js");
-        expect(diffData.diff).toEqual("");
+        expect(assertNewFileEvent(repoPath, newRelPath)).toBe(true);
     });
 
     test("Valid File",  async () => {
         const handler = new eventHandler();
         handler.handleNewFile(newFilePath);
-        // Verify file has been created in the .shadow repo and .originals repos
-        expect(fs.existsSync(shadowFilePath)).toBe(true);
-        expect(fs.existsSync(originalsFilePath)).toBe(true);
-        // Verify correct diff file has been generated
-        let diffFiles = fs.readdirSync(diffsRepo);
-        expect(diffFiles).toHaveLength(1);
-        const diffFilePath = path.join(diffsRepo, diffFiles[0]);
-        const diffData = readYML(diffFilePath);
-        expect(diffData.source).toEqual(DIFF_SOURCE);
-        expect(diffData.is_new_file).toBe(true);
-        expect(diffData.is_rename).toBeFalsy();
-        expect(diffData.is_deleted).toBeFalsy();
-        expect(diffData.repo_path).toEqual(repoPath);
-        expect(diffData.branch).toEqual(DEFAULT_BRANCH);
-        expect(diffData.file_relative_path).toEqual("new.js");
-        expect(diffData.diff).toEqual("");
+        expect(assertNewFileEvent(repoPath, newRelPath)).toBe(true);
+    });
+
+    test("With Daemon: Valid File",  async () => {
+        const handler = new eventHandler();
+        handler.handleNewFile(newFilePath);
+        await populateBuffer();
+        expect(assertNewFileEvent(repoPath, newRelPath)).toBe(true);
     });
 
     test("with syncignored file",  async () => {