+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.29.22] - 2021-10-22
 ### Updated
 - Using random ports for Auth Server
+- Updated events-tests to run in parallel with Daemon
+- Sending IDE and Platform info to server for repo details and diffs.
 
 ## [2.29.21] - 2021-10-21
 ### Fixed